add tests for WS
